Added docblocks and moved debug stuff to separate function

+++ classes/core/skynet.php
@@ -6,41 +6,54 @@
 
 class Core_Skynet {
     
-    /*
-     * __construct()
-     */
+    /**
+     * $reg
+     * Link to registery
+     * @access private
+    */
+    private $reg = null;
     
-    private $reg;
-    
+    /**
+     * __construct
+     * creates base for CMS
+    */
     function __construct() {
-        $this->init();
+        $this->initBaseClasses();
+        $this->testBaseClasses();
     }
     
-    private function init(){
+    /**
+     * initBaseClasses
+     * Initializes most basic classes
+     * @access private
+    */
+    private function initBaseClasses(){
         $this->reg = new Core_Registery();
         new Core_Config();
         new Core_Database();
-        
-        
-        /**
-         * Debug lines
-        */
-        var_export($this->reg->database->insert("testje",array("id"=>1)));
-        echo PHP_EOL;
-        var_export($this->reg->database->insert("testje",array("id"=>2)));
-        echo PHP_EOL;
-        var_export($this->reg->database->insert("testje",array("id"=>3)));
-        echo PHP_EOL;
-        print_r($this->reg->database->update("testje",array("id"=>5),array("id"=>3)));
-        echo PHP_EOL;
-        print_r($this->reg->database->delete("testje",array("id"=>2)));
-        echo PHP_EOL;
-        print_r($this->reg->database->select("testje","*"));
-        echo PHP_EOL;
-        var_export($this->reg->database->clearTable("testje"));
-        /**
-         * End of debug lines
-        */
+    }
+    
+    /**
+     * 
+    */
+    private function testBaseClasses(){
+        //test config & registery
+        if($this->reg->config->getvar("cms_debug")){
+            //test database
+            var_export($this->reg->database->insert("testje",array("id"=>1)));
+            echo PHP_EOL;
+            var_export($this->reg->database->insert("testje",array("id"=>2)));
+            echo PHP_EOL;
+            var_export($this->reg->database->insert("testje",array("id"=>3)));
+            echo PHP_EOL;
+            print_r($this->reg->database->update("testje",array("id"=>5),array("id"=>3)));
+            echo PHP_EOL;
+            print_r($this->reg->database->delete("testje",array("id"=>2)));
+            echo PHP_EOL;
+            print_r($this->reg->database->select("testje","*"));
+            echo PHP_EOL;
+            var_export($this->reg->database->clearTable("testje"));
+        }
     }
 }
